Enemy and Wave Managers
Finished enemy and wave managers so the game is now functional and the player can progress through infinite waves of enemies. Currently, since there are 4 types of enemies, the player will progress through 25 waves of increasing difficulty before all enemies spawned will be random. All that is left to do now is fix the settings and player death screens.